+++ sdk/crypto/basic/crypto_keychain.hpp
@@ -523,6 +523,15 @@ class crypto_keychain {
     return_t add_mlkem_priv_b16(crypto_key* cryptokey, uint32 nid, const char* keypair, key_encoding_t encoding, const keydesc& desc);
     return_t add_mlkem_priv_b16rfc(crypto_key* cryptokey, uint32 nid, const char* keypair, key_encoding_t encoding, const keydesc& desc);
 
+    /**
+     * @brief MLDSA
+     * @param uint32 nid [in] NID_ML_DSA_44, NID_ML_DSA_65, NID_ML_DSA_87
+     * @remarks openssl-3.5 required
+     */
+    return_t add_mldsa(crypto_key* cryptokey, uint32 nid, const keydesc& desc);
+    return_t add_mldsa_pub(crypto_key* cryptokey, uint32 nid, const byte_t* pub, size_t pubsize, key_encoding_t encoding, const keydesc& desc);
+    return_t add_mldsa_priv(crypto_key* cryptokey, uint32 nid, const binary_t& keypair, key_encoding_t encoding, const keydesc& desc);
+
    protected:
 };