+++ sdk/crypto/basic/crypto_sign.hpp
@@ -185,6 +185,16 @@ class crypto_sign_dsa : public crypto_sign {
     virtual return_t verify(const EVP_PKEY* pkey, const binary_t& input, const binary_t& signature, uint32 flags = 0);
 };
 
+class crypto_sign_mldsa : public crypto_sign {
+   public:
+    crypto_sign_mldsa();
+
+    virtual return_t sign(const EVP_PKEY* pkey, const byte_t* stream, size_t size, binary_t& signature, uint32 flags = 0);
+    virtual return_t verify(const EVP_PKEY* pkey, const byte_t* stream, size_t size, const binary_t& signature, uint32 flags = 0);
+    virtual return_t sign(const EVP_PKEY* pkey, const binary_t& input, binary_t& signature, uint32 flags = 0);
+    virtual return_t verify(const EVP_PKEY* pkey, const binary_t& input, const binary_t& signature, uint32 flags = 0);
+};
+
 // not implemented
 // class crypto_sign_rsa_x931 : public crypto_sign {};