+++ sdk/crypto/basic/sign/openssl_sign_mldsa.cpp
@@ -1,0 +1,125 @@
+/* vim: set tabstop=4 shiftwidth=4 softtabstop=4 expandtab smarttab : */
+/**
+ * @file {file}
+ * @author Soo Han, Kim (princeb612.kr@gmail.com)
+ * @desc
+ *
+ * Revision History
+ * Date         Name                Description
+ */
+
+#include <hotplace/sdk/crypto/basic/crypto_advisor.hpp>
+#include <hotplace/sdk/crypto/basic/evp_pkey.hpp>
+#include <hotplace/sdk/crypto/basic/openssl_hash.hpp>
+#include <hotplace/sdk/crypto/basic/openssl_sdk.hpp>
+#include <hotplace/sdk/crypto/basic/openssl_sign.hpp>
+
+namespace hotplace {
+namespace crypto {
+
+return_t openssl_sign::sign_mldsa(const EVP_PKEY* pkey, const binary_t& input, binary_t& signature, uint32 flags) {
+    return sign_mldsa(pkey, input.data(), input.size(), signature, flags);
+}
+
+return_t openssl_sign::sign_mldsa(const EVP_PKEY* pkey, const byte_t* stream, size_t size, binary_t& signature, uint32 flags) {
+    return_t ret = errorcode_t::success;
+    EVP_MD_CTX* ctx = nullptr;
+    int ret_test = 0;
+
+    __try2 {
+        signature.clear();
+
+        if (nullptr == pkey || nullptr == stream) {
+            ret = errorcode_t::invalid_parameter;
+            __leave2;
+        }
+
+        auto kty = ktyof_evp_pkey(pkey);
+        if (kty_mldsa != kty) {
+            ret = errorcode_t::invalid_context;
+            __leave2;
+        }
+
+        ctx = EVP_MD_CTX_new();
+        ret_test = EVP_DigestSignInit(ctx, nullptr, nullptr, nullptr, (EVP_PKEY*)pkey);
+        if (1 != ret_test) {
+            ret = errorcode_t::internal_error;
+            __leave2_trace_openssl(ret);
+        }
+
+        /**
+         * ML-DSA-44 2420
+         * ML-DSA-65 3309
+         * ML-DSA-87 4627
+         */
+        size_t dgstsize = 0;
+        ret_test = EVP_DigestSign(ctx, nullptr, &dgstsize, stream, size);
+        if (1 != ret_test) {
+            ret = errorcode_t::internal_error;
+            __leave2_trace_openssl(ret);
+        }
+        signature.resize(dgstsize);
+
+        ret_test = EVP_DigestSign(ctx, signature.data(), &dgstsize, stream, size);
+        if (1 != ret_test) {
+            ret = errorcode_t::internal_error;
+            __leave2_trace_openssl(ret);
+        }
+        signature.resize(dgstsize);
+    }
+    __finally2 {
+        if (ctx) {
+            EVP_MD_CTX_destroy(ctx);
+        }
+    }
+    return ret;
+}
+
+return_t openssl_sign::verify_mldsa(const EVP_PKEY* pkey, const binary_t& input, const binary_t& signature, uint32 flags) {
+    return verify_mldsa(pkey, input.data(), input.size(), signature, flags);
+}
+
+return_t openssl_sign::verify_mldsa(const EVP_PKEY* pkey, const byte_t* stream, size_t size, const binary_t& signature, uint32 flags) {
+    return_t ret = errorcode_t::success;
+    EVP_MD_CTX* ctx = nullptr;
+    int ret_test = 0;
+
+    __try2 {
+        if (nullptr == pkey || nullptr == stream) {
+            ret = errorcode_t::invalid_parameter;
+            __leave2;
+        }
+
+        auto kty = ktyof_evp_pkey(pkey);
+        if (kty_mldsa != kty) {
+            ret = errorcode_t::invalid_context;
+            __leave2;
+        }
+
+        ret = errorcode_t::error_verify;
+
+        ctx = EVP_MD_CTX_new();
+        ret_test = EVP_DigestVerifyInit(ctx, nullptr, nullptr, nullptr, (EVP_PKEY*)pkey);
+        if (1 != ret_test) {
+            ret = errorcode_t::internal_error;
+            __leave2_trace_openssl(ret);
+        }
+
+        ret_test = EVP_DigestVerify(ctx, signature.data(), signature.size(), stream, size);
+        if (1 != ret_test) {
+            ret = errorcode_t::error_verify;
+            __leave2_trace_openssl(ret);
+        }
+
+        ret = errorcode_t::success;
+    }
+    __finally2 {
+        if (ctx) {
+            EVP_MD_CTX_destroy(ctx);
+        }
+    }
+    return ret;
+}
+
+}  // namespace crypto
+}  // namespace hotplace