+++ testapplet/httpserver1/sample.hpp
@@ -13,6 +13,11 @@
 #include <hotplace/sdk/sdk.hpp>
 #include <hotplace/testcase/test.hpp>
 
+enum option_flag_t {
+    option_flag_cert_rsa = (1 << 4),
+    option_flag_cert_ecdsa = (1 << 5),
+    option_flag_cert_mldsa = (1 << 6),
+};
 struct OPTION : public CMDLINEOPTION {
     int run;
     int port;
@@ -20,9 +25,10 @@ struct OPTION : public CMDLINEOPTION {
     int content_encoding;
     int trial;
     int keylog;
+    uint32 flags;
     std::string cs;
 
-    OPTION() : CMDLINEOPTION(), run(0), port(8080), port_tls(9000), content_encoding(0), trial(0), keylog(0) {}
+    OPTION() : CMDLINEOPTION(), run(0), port(8080), port_tls(9000), content_encoding(0), trial(0), keylog(0), flags(0) {}
 };
 
 extern test_case _test_case;