+++ testcase/crypto/sample.hpp
@@ -82,6 +82,7 @@ void testcase_crypto_sign();
 void testcase_dsa();
 void testcase_ecdsa();
 void testcase_hmac();
+void testcase_mldsa();
 void testcase_rsassa();
 void testcase_x509();